+++ docs/metro/js/metro.js
@@ -119,7 +119,7 @@ var isTouch = (('ontouchstart' in window) || (navigator.MaxTouchPoints > 0) || (
 var Metro = {
 
     version: "4.2.43",
-    compileTime: "15/05/2019 22:12:46",
+    compileTime: "16/05/2019 11:00:17",
     buildNumber: "724",
     isTouchable: isTouch,
     fullScreenEnabled: document.fullscreenEnabled,
@@ -19478,6 +19478,7 @@ var Streamer = {
         }
 
         element.data("stream", -1);
+        element.find(".events-area").scrollLeft(0);
 
         this._createEvents();
 
@@ -19495,14 +19496,20 @@ var Streamer = {
 
     _fireScroll: function(){
         var that = this, element = this.element, o = this.options;
+        var scrollable = element.find(".events-area");
         var oldScroll = this.scroll;
-        this.scrollDir = this.scroll < element[0].scrollLeft ? "left" : "right";
-        this.scroll = element[0].scrollLeft;
 
-        Utils.exec(o.onEventsScroll, [element[0].scrollLeft, oldScroll, this.scrollDir], element[0]);
+        if (scrollable.length === 0) {
+            return undefined;
+        }
+
+        this.scrollDir = this.scroll < scrollable[0].scrollLeft ? "left" : "right";
+        this.scroll = scrollable[0].scrollLeft;
+
+        Utils.exec(o.onEventsScroll, [scrollable[0].scrollLeft, oldScroll, this.scrollDir], element[0]);
 
         element.fire("eventsscroll", {
-            scrollLeft: element[0].scrollLeft,
+            scrollLeft: scrollable[0].scrollLeft,
             oldScroll: oldScroll,
             scrollDir: that.scrollDir
         });