+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 + [x] Utils: rename method `isJQueryObject` to `isJQuery`
 + [x] Utils: add methods `isM4Q`, `isQ`. Method `isQ` return one of `isJQuery` or `isM4Q`
 + [x] Streamer: fix scrolling with `apple magic mouse` and `firefox`
-
++ [x] Streamer: fix scroll position for `events-area` when source changed
+ 
 ### 4.2.42
 + [x] General: improved components initialization
 + [x] Hotkeys: now you can add or change `hotkey` at runtime